ref: generate fack id

+++ src/ShoppingCardApi/UseCases/ShoppingCart/RemoveProductFromCart/Command/RemoveProductFromCartCommandHandler.cs
@@ -1,18 +1,30 @@
 using MediatR;
+using ShoppingCardApi.Contracts;
 using ShoppingCardApi.Domain;
 
 namespace ShoppingCardApi.UseCases.ShoppingCart.RemoveProductFromCart;
 
-public class RemoveProductFromCartCommandHandler:IRequestHandler<RemoveProductFromCart>
+public class RemoveProductFromCartCommandHandler : IRequestHandler<RemoveProductFromCart, Result>
 {
     private readonly IRepository<Domain.ShoppingCart> _repository;
     public RemoveProductFromCartCommandHandler(IRepository<Domain.ShoppingCart> repository)
     {
         _repository = repository;
     }
-    public async Task Handle(RemoveProductFromCart request, CancellationToken cancellationToken)
+    public async Task<Result> Handle(RemoveProductFromCart request, CancellationToken cancellationToken)
     {
-      var getProduct=  await _repository.GetAsync(request.id);
-      await _repository.DeleteAsync(getProduct);
+        try
+        {
+            var getProduct = await _repository.GetAsync(request.id);
+            if (getProduct == null)
+                return Result.Failure();
+            await _repository.DeleteAsync(getProduct);
+            return Result.Success();
+        }
+        catch (Exception ex)
+        {
+            return Result.Failure();
+        }
+
     }
 }
